+++ src/compiler/elab_core.c
@@ -2046,6 +2046,8 @@ void elab_init_state(Elab *e, Arena *arena, SymbolTable *st) {
     e->sym_caret_relevant  = intern_cstr(st, "^relevant");
     e->sym_caret_atomic   = intern_cstr(st, "^atomic");   /* G4a */
     e->sym_caret_thread_local = intern_cstr(st, "^thread-local"); /* G4b */
+    e->sym_caret_deferred_init = intern_cstr(st, "^deferred-init");
+    e->sym_deferred_init       = intern_cstr(st, "__tur-deferred-init__");
     /* LB1: ^borrow -- non-consuming linear/affine handle accessor parameter */
     e->sym_caret_borrow    = intern_cstr(st, "^borrow");
     e->sym_caret_fat       = intern_cstr(st, "^fat");  /* A#1: fat-closure param */